2002-10-28  Tim Coleman (tim@timcoleman.com)         * Tds.cs:                 Add much handling to get the following types                 working in queries: string, int, decimal                 All sorts of other bug fixing and general                 purpose hackery.         * Tds42.cs:         * Tds50.cs:         * Tds70.cs:                 ProcessColumnInfo was abstracted in Tds.cs                 and subclassed, because 7.0 provides the information                 in a different format.         * TdsColumnSchema.cs:                 Added size, precision, scale         * TdsComm.cs:                 Fixed GetString because it wasn't working properly                 with TDS 7.0.         * TdsMessage.cs:                 Change state and severity to byte instead of int                 because that is more consitent.
svn path=/trunk/mcs/; revision=8645

+++ mcs/class/Mono.Data.Tds/Mono.Data.Tds.Protocol/TdsMessage.cs
@@ -16,8 +16,8 @@ namespace Mono.Data.TdsClient.Internal {
 		#region Fields
 
 		int number = 0;
-		int state = 0;
-		int severity = 0;
+		byte state = 0;
+		byte severity = 0;
 		string message = String.Empty;
 		string server = String.Empty;
 		string procName = String.Empty;
@@ -47,7 +47,7 @@ namespace Mono.Data.TdsClient.Internal {
 			set { procName = value; }
 		}
 
-		public int State {
+		public byte State {
 			get { return state; }
 			set { state = value; }
 		}
@@ -57,7 +57,7 @@ namespace Mono.Data.TdsClient.Internal {
 			set { server = value; }
 		}
 
-		public int Severity {
+		public byte Severity {
 			get { return severity; }
 			set { severity = value; }
 		}
